net: dsa: ksz8873: fix mismatched return value
Both read and write operations for the regmap are supposed to return an
int, not a ssize_t and build warns about this. Fix it.

Link: https://lore.barebox.org/20230125075354.1760273-2-a.fatoum@pengutronix.de

+++ drivers/net/ksz8873.c
@@ -96,7 +96,7 @@ static int ksz8873_mdio_read(void *ctx, unsigned int reg, unsigned int *val)
 	return 0;
 }
 
-static ssize_t ksz8873_mdio_write(void *ctx, unsigned int reg, unsigned int val)
+static int ksz8873_mdio_write(void *ctx, unsigned int reg, unsigned int val)
 {
 	struct ksz8873_switch *priv = ctx;
 	struct phy_device *mdiodev = priv->mdiodev;
